test: refactor and merge `no_ssr` and `different_domains` tests
