+++ specs/different_domains/different_domains_multi_locales_prefix_except_default.spec.ts
@@ -1,13 +1,14 @@
 import { test, expect, describe } from 'vitest'
 import { fileURLToPath } from 'node:url'
-import { setup, $fetch, undiciRequest } from './utils'
-import { getDom } from './helper'
+import { setup, $fetch, undiciRequest } from '../utils'
+import { getDom } from '../helper'
 
 await setup({
-  rootDir: fileURLToPath(new URL(`./fixtures/different_domains`, import.meta.url)),
+  rootDir: fileURLToPath(new URL(`../fixtures/different_domains`, import.meta.url)),
   // overrides
   nuxtConfig: {
     i18n: {
+      baseUrl: 'http://localhost:3000',
       locales: [
         {
           code: 'en',
@@ -26,11 +27,19 @@ await setup({
           code: 'fr',
           iso: 'fr-FR',
           name: 'Français',
-          domain: 'fr.nuxt-app.localhost'
+          domain: 'fr.nuxt-app.localhost',
+          domainDefault: true
+        },
+        {
+          code: 'ja',
+          iso: 'jp-JA',
+          name: 'Japan',
+          domain: 'ja.nuxt-app.localhost',
+          domainDefault: true
         }
       ],
       differentDomains: true,
-      strategy: 'prefix',
+      strategy: 'prefix_except_default',
       detectBrowserLanguage: {
         useCookie: true
       }
@@ -40,10 +49,11 @@ await setup({
 
 describe('detection locale with host on server', () => {
   test.each([
-    ['en', 'nuxt-app.localhost', 'Homepage'],
-    ['fr', 'fr.nuxt-app.localhost', 'Accueil']
-  ])('%s host', async (locale, host, header) => {
-    const res = await undiciRequest('/' + locale, {
+    ['/', 'en', 'nuxt-app.localhost', 'Homepage'],
+    ['/no', 'no', 'nuxt-app.localhost', 'Hjemmeside'],
+    ['/', 'fr', 'fr.nuxt-app.localhost', 'Accueil']
+  ])('%s host', async (path, locale, host, header) => {
+    const res = await undiciRequest(path, {
       headers: {
         Host: host
       }
@@ -56,7 +66,7 @@ describe('detection locale with host on server', () => {
 })
 
 test('detection locale with x-forwarded-host on server', async () => {
-  const html = await $fetch('/fr', {
+  const html = await $fetch('/', {
     headers: {
       'X-Forwarded-Host': 'fr.nuxt-app.localhost'
     }
@@ -68,19 +78,19 @@ test('detection locale with x-forwarded-host on server', async () => {
 })
 
 test('pass `<NuxtLink> to props', async () => {
-  const res = await undiciRequest('/fr', {
+  const res = await undiciRequest('/', {
     headers: {
       Host: 'fr.nuxt-app.localhost'
     }
   })
   const dom = getDom(await res.body.text())
   expect(dom.querySelector('#switch-locale-path-usages .switch-to-en a').getAttribute('href')).toEqual(
-    `http://nuxt-app.localhost/en`
+    `http://nuxt-app.localhost`
   )
   expect(dom.querySelector('#switch-locale-path-usages .switch-to-no a').getAttribute('href')).toEqual(
     `http://nuxt-app.localhost/no`
   )
   expect(dom.querySelector('#switch-locale-path-usages .switch-to-fr a').getAttribute('href')).toEqual(
-    `http://fr.nuxt-app.localhost/fr`
+    `http://fr.nuxt-app.localhost`
   )
 })
