--- specs/fixtures/no_ssr/app.vue
@@ -1,5 +1,0 @@
-<template>
-  <div>
-    <NuxtPage />
-  </div>
-</template>
