--- specs/fixtures/no_ssr/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "nuxt3-test-issues-2313",
-  "private": true,
-  "scripts": {
-    "build": "nuxt build",
-    "dev": "nuxt dev",
-    "generate": "nuxt generate",
-    "preview": "nuxt preview"
-  },
-  "devDependencies": {
-    "@nuxtjs/i18n": "latest",
-    "nuxt": "latest"
-  }
-}
