--- specs/fixtures/no_ssr/pages/index.vue
@@ -1,10 +1,0 @@
-<template>
-  <div>
-    <div id="top">Foo</div>
-    Locale: <code id="locale">{{ locale }}</code>
-  </div>
-</template>
-
-<script setup lang="ts">
-const { locale } = useI18n()
-</script>
